Move KeyActionMap to new KeyboardShortcutsComponent
